Update tests and move them to distinct files
