+++ pkg/model/model_test.go
@@ -3,11 +3,9 @@ package model
 import (
 	"math"
 	"reflect"
-	"strconv"
 	"testing"
 	"time"
 
-	"github.com/akosgarai/playground_engine/pkg/material"
 	"github.com/akosgarai/playground_engine/pkg/mesh"
 	"github.com/akosgarai/playground_engine/pkg/primitives/boundingobject"
 	"github.com/akosgarai/playground_engine/pkg/testhelper"
@@ -957,1550 +955,5 @@ func TestTerrainBuilderSetLiquidDetailMultiplier(t *testing.T) {
 	terr.SetLiquidDetailMultiplier(dm)
 	if terr.liquidDetailMultiplier != dm {
 		t.Errorf("Invalid liquid detail multiplier. Instead of '%d', we have '%d'.", dm, terr.liquidDetailMultiplier)
-	}
-}
-func testFormItemBool(t *testing.T) *FormItemBool {
-	mat := material.Chrome
-	pos := mgl32.Vec3{0, 0, 0}
-	fi := NewFormItemBool(DefaultMaxWidth, ITEM_WIDTH_HALF, DefaultFormItemLabel, DefaultFormItemDescription, mat, pos, wrapperMock)
-
-	if fi.label != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.label)
-	}
-	if fi.description != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.description)
-	}
-	return fi
-}
-func TestNewFormItemBool(t *testing.T) {
-	_ = testFormItemBool(t)
-}
-func TestFormItemBoolGetLabel(t *testing.T) {
-	fi := testFormItemBool(t)
-	if fi.GetLabel() != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.GetLabel())
-	}
-}
-func TestFormItemBoolGetDescription(t *testing.T) {
-	fi := testFormItemBool(t)
-	if fi.GetDescription() != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.GetDescription())
-	}
-}
-func TestFormItemBoolGetValue(t *testing.T) {
-	fi := testFormItemBool(t)
-	val := true
-	fi.value = val
-	if fi.GetValue() != val {
-		t.Errorf("Invalid form item value. Instead of '%v', it is '%v'.", val, fi.GetValue())
-	}
-}
-func TestFormItemBoolSetValue(t *testing.T) {
-	fi := testFormItemBool(t)
-	val := true
-	fi.value = val
-	fi.SetValue(!val)
-	if fi.GetValue() != !val {
-		t.Errorf("Invalid form item value. Instead of '%v', it is '%v'.", !val, fi.GetValue())
-	}
-}
-func TestFormItemBoolGetSurface(t *testing.T) {
-	fi := testFormItemBool(t)
-	if fi.GetSurface() != fi.meshes[0] {
-		t.Error("Invalid surface mesh")
-	}
-}
-func TestFormItemBoolGetLight(t *testing.T) {
-	fi := testFormItemBool(t)
-	if fi.GetLight() != fi.meshes[1] {
-		t.Error("Invalid light mesh")
-	}
-}
-func TestFormItemBoolValueToString(t *testing.T) {
-	fi := testFormItemBool(t)
-	if fi.ValueToString() != "false" {
-		t.Error("Invalid value string.")
-	}
-	fi.SetValue(true)
-	if fi.ValueToString() != "true" {
-		t.Error("Invalid value string.")
-	}
-}
-func testFormItemInt(t *testing.T) *FormItemInt {
-	mat := material.Chrome
-	pos := mgl32.Vec3{0, 0, 0}
-	fi := NewFormItemInt(DefaultMaxWidth, ITEM_WIDTH_HALF, DefaultFormItemLabel, DefaultFormItemDescription, mat, pos, wrapperMock)
-
-	if fi.label != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.label)
-	}
-	if fi.description != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.description)
-	}
-	return fi
-}
-func TestNewFormItemInt(t *testing.T) {
-	_ = testFormItemInt(t)
-}
-func TestFormItemIntGetLabel(t *testing.T) {
-	fi := testFormItemInt(t)
-	if fi.GetLabel() != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.GetLabel())
-	}
-}
-func TestFormItemIntGetDescription(t *testing.T) {
-	fi := testFormItemInt(t)
-	if fi.GetDescription() != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.GetDescription())
-	}
-}
-func TestFormItemIntGetValue(t *testing.T) {
-	fi := testFormItemInt(t)
-	val := 3
-	fi.value = "3"
-	if fi.GetValue() != val {
-		t.Errorf("Invalid form item value. Instead of '%d', it is '%d'.", val, fi.GetValue())
-	}
-}
-func TestFormItemIntGetSurface(t *testing.T) {
-	fi := testFormItemInt(t)
-	if fi.GetSurface() != fi.meshes[0] {
-		t.Error("Invalid surface mesh")
-	}
-}
-func TestFormItemIntGetTarget(t *testing.T) {
-	fi := testFormItemInt(t)
-	if fi.GetTarget() != fi.meshes[1] {
-		t.Error("Invalid target mesh")
-	}
-}
-func TestFormItemIntAddCursor(t *testing.T) {
-	fi := testFormItemInt(t)
-	if len(fi.meshes) != 2 {
-		t.Error("Invalid number of target mesh")
-	}
-	fi.AddCursor()
-	if len(fi.meshes) != 3 {
-		t.Error("Invalid number of target mesh")
-	}
-}
-func TestFormItemIntDeleteCursor(t *testing.T) {
-	fi := testFormItemInt(t)
-	fi.AddCursor()
-	if len(fi.meshes) != 3 {
-		t.Error("Invalid number of target mesh")
-	}
-	fi.DeleteCursor()
-	if len(fi.meshes) != 2 {
-		t.Error("Invalid number of target mesh")
-	}
-}
-func TestFormItemIntCharCallback(t *testing.T) {
-	fi := testFormItemInt(t)
-	fi.AddCursor()
-	fi.DeleteLastCharacter()
-	fi.CharCallback('b', 0.1)
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.value != "0" {
-		t.Errorf("Invalid value. Instead of '0', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	fi.CharCallback('-', 0.1)
-	if fi.value != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "-1" {
-		t.Errorf("Invalid value. Instead of '-1', we have '%s'.", fi.value)
-	}
-	fi = testFormItemInt(t)
-	fi.AddCursor()
-	fi.CharCallback('1', 0.1)
-	if fi.value != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	fi.value = strconv.Itoa(math.MaxInt32 - 10)
-	fi.CharCallback('1', 0.1)
-	if fi.value != strconv.Itoa(math.MaxInt32-10) {
-		t.Errorf("Invalid value. Instead of '%d', we have '%s'.", math.MaxInt32-10, fi.value)
-	}
-	fi = testFormItemInt(t)
-	fi.AddCursor()
-	validator := func(i int) bool {
-		return i <= 100 && i >= -10
-	}
-	fi.SetValidator(validator)
-	fi.CharCallback('1', 0.1)
-	if fi.value != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "11" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "11" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	fi.DeleteLastCharacter()
-	fi.CharCallback('-', 0.1)
-	fi.CharCallback('1', 0.1)
-	if fi.value != "-1" {
-		t.Errorf("Invalid value. Instead of '-1', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "-1" {
-		t.Errorf("Invalid value. Instead of '-1', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.value != "-10" {
-		t.Errorf("Invalid value. Instead of '-10', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.value != "-10" {
-		t.Errorf("Invalid value. Instead of '-10', we have '%s'.", fi.value)
-	}
-}
-func TestFormItemIntValueToString(t *testing.T) {
-	fi := testFormItemInt(t)
-	val := fi.ValueToString()
-	if val != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", val)
-	}
-	fi.value = strconv.Itoa(3)
-	val = fi.ValueToString()
-	if val != "3" {
-		t.Errorf("Invalid value. Instead of '3', we have '%s'.", val)
-	}
-}
-func TestFormItemIntDeleteLastCharacter(t *testing.T) {
-	fi := testFormItemInt(t)
-	fi.value = "12345"
-	fi.typeState = "PI"
-	fi.charOffsets = []float32{0.1, 0.1, 0.1, 0.1, 0.1}
-	fi.DeleteLastCharacter()
-	if fi.value != "1234" {
-		t.Errorf("Invalid value. Instead of '1234', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "123" {
-		t.Errorf("Invalid value. Instead of '123', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "12" {
-		t.Errorf("Invalid value. Instead of '12', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-	fi.value = "-1234"
-	fi.typeState = "NI"
-	fi.charOffsets = []float32{0.1, 0.1, 0.1, 0.1, 0.1}
-	fi.DeleteLastCharacter()
-	if fi.value != "-123" {
-		t.Errorf("Invalid value. Instead of '-123', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "-12" {
-		t.Errorf("Invalid value. Instead of '-12', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "-1" {
-		t.Errorf("Invalid value. Instead of '-1', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-}
-func testFormItemFloat(t *testing.T) *FormItemFloat {
-	mat := material.Chrome
-	pos := mgl32.Vec3{0, 0, 0}
-	fi := NewFormItemFloat(DefaultMaxWidth, ITEM_WIDTH_HALF, DefaultFormItemLabel, DefaultFormItemDescription, mat, pos, wrapperMock)
-
-	if fi.label != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.label)
-	}
-	if fi.description != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.description)
-	}
-	return fi
-}
-func TestNewFormItemFloat(t *testing.T) {
-	_ = testFormItemFloat(t)
-}
-func TestFormItemFloatGetLabel(t *testing.T) {
-	fi := testFormItemFloat(t)
-	if fi.GetLabel() != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.GetLabel())
-	}
-}
-func TestFormItemFloatGetDescription(t *testing.T) {
-	fi := testFormItemFloat(t)
-	if fi.GetDescription() != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.GetDescription())
-	}
-}
-func TestFormItemFloatGetValue(t *testing.T) {
-	fi := testFormItemFloat(t)
-	valI := 3
-	fi.value = "3"
-	if fi.GetValue() != float32(valI) {
-		t.Errorf("Invalid form item value. Instead of '%d', it is '%f'.", valI, fi.GetValue())
-	}
-	fi.value = "3.2"
-	if fi.GetValue() != 3.2 {
-		t.Errorf("Invalid form item value. Instead of '3.2', it is '%f'.", fi.GetValue())
-	}
-}
-func TestFormItemFloatGetSurface(t *testing.T) {
-	fi := testFormItemFloat(t)
-	if fi.GetSurface() != fi.meshes[0] {
-		t.Error("Invalid surface mesh")
-	}
-}
-func TestFormItemFloatGetTarget(t *testing.T) {
-	fi := testFormItemFloat(t)
-	if fi.GetTarget() != fi.meshes[1] {
-		t.Error("Invalid target mesh")
-	}
-}
-func TestFormItemFloatAddCursor(t *testing.T) {
-	fi := testFormItemFloat(t)
-	if len(fi.meshes) != 2 {
-		t.Error("Invalid number of target mesh")
-	}
-	fi.AddCursor()
-	if len(fi.meshes) != 3 {
-		t.Error("Invalid number of target mesh")
-	}
-}
-func TestFormItemFloatDeleteCursor(t *testing.T) {
-	fi := testFormItemFloat(t)
-	fi.AddCursor()
-	if len(fi.meshes) != 3 {
-		t.Error("Invalid number of target mesh")
-	}
-	fi.DeleteCursor()
-	if len(fi.meshes) != 2 {
-		t.Error("Invalid number of target mesh")
-	}
-}
-func TestFormItemFloatCharCallback(t *testing.T) {
-	fi := testFormItemFloat(t)
-	fi.AddCursor()
-	// start with 0
-	fi.CharCallback('0', 0.1)
-	if fi.value != "0" {
-		t.Errorf("Invalid value. Instead of '0', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "P0" {
-		t.Errorf("Invalid typeState. Instead of 'P0', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "0" {
-		t.Errorf("Invalid value. Instead of '0', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "P0" {
-		t.Errorf("Invalid typeState. Instead of 'P0', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.value != "0." {
-		t.Errorf("Invalid value. Instead of '0.', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "P." {
-		t.Errorf("Invalid typeState. Instead of 'P.', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('2', 0.1)
-	if fi.value != "0.2" {
-		t.Errorf("Invalid value. Instead of '0.2', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "PF" {
-		t.Errorf("Invalid typeState. Instead of 'PF', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('8', 0.1)
-	if fi.value != "0.28" {
-		t.Errorf("Invalid value. Instead of '0.28', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "PF" {
-		t.Errorf("Invalid typeState. Instead of 'PF', we have '%s'.", fi.typeState)
-	}
-	// start with .
-	fi = testFormItemFloat(t)
-	fi.AddCursor()
-	fi.CharCallback('.', 0.1)
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "PI" {
-		t.Errorf("Invalid typeState. Instead of 'PI', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('2', 0.1)
-	if fi.value != "12" {
-		t.Errorf("Invalid value. Instead of '12', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "PI" {
-		t.Errorf("Invalid typeState. Instead of 'PI', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.value != "12." {
-		t.Errorf("Invalid value. Instead of '12.', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "P." {
-		t.Errorf("Invalid typeState. Instead of 'P.', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('8', 0.1)
-	if fi.value != "12.8" {
-		t.Errorf("Invalid value. Instead of '12.8', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "PF" {
-		t.Errorf("Invalid typeState. Instead of 'PF', we have '%s'.", fi.typeState)
-	}
-	// start with -
-	fi = testFormItemFloat(t)
-	fi.AddCursor()
-	fi.CharCallback('-', 0.1)
-	if fi.value != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "N" {
-		t.Errorf("Invalid typeState. Instead of 'N', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.value != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "N" {
-		t.Errorf("Invalid typeState. Instead of 'N', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "-1" {
-		t.Errorf("Invalid value. Instead of '-1', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "NI" {
-		t.Errorf("Invalid typeState. Instead of 'NI', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('2', 0.1)
-	if fi.value != "-12" {
-		t.Errorf("Invalid value. Instead of '-12', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "NI" {
-		t.Errorf("Invalid typeState. Instead of 'NI', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.value != "-12." {
-		t.Errorf("Invalid value. Instead of '-12.', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "N." {
-		t.Errorf("Invalid typeState. Instead of 'N.', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('8', 0.1)
-	if fi.value != "-12.8" {
-		t.Errorf("Invalid value. Instead of '-12.8', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "NF" {
-		t.Errorf("Invalid typeState. Instead of 'NF', we have '%s'.", fi.typeState)
-	}
-	// start with -0.
-	fi = testFormItemFloat(t)
-	fi.AddCursor()
-	fi.CharCallback('-', 0.1)
-	if fi.value != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "N" {
-		t.Errorf("Invalid typeState. Instead of 'N', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.value != "-0" {
-		t.Errorf("Invalid value. Instead of '-0', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "N0" {
-		t.Errorf("Invalid typeState. Instead of 'N0', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.value != "-0." {
-		t.Errorf("Invalid value. Instead of '-0.', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "N." {
-		t.Errorf("Invalid typeState. Instead of 'N.', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('8', 0.1)
-	if fi.value != "-0.8" {
-		t.Errorf("Invalid value. Instead of '-0.8', we have '%s'.", fi.value)
-	}
-	if fi.typeState != "NF" {
-		t.Errorf("Invalid typeState. Instead of 'NF', we have '%s'.", fi.typeState)
-	}
-	// validator
-	fi = testFormItemFloat(t)
-	fi.AddCursor()
-	validator := func(i float32) bool {
-		return i <= 1.0 && i >= 0.0
-	}
-	fi.SetValidator(validator)
-	fi.CharCallback('-', 0.1)
-	if fi.value != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.value != "-0" {
-		t.Errorf("Invalid value. Instead of '-0', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.value != "-0." {
-		t.Errorf("Invalid value. Instead of '-0.', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('3', 0.1)
-	if fi.value != "-0." {
-		t.Errorf("Invalid value. Instead of '-0.', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	fi.DeleteLastCharacter()
-	fi.DeleteLastCharacter()
-	fi.CharCallback('1', 0.1)
-	if fi.value != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.value != "1." {
-		t.Errorf("Invalid value. Instead of '1.', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('3', 0.1)
-	if fi.value != "1." {
-		t.Errorf("Invalid value. Instead of '1.', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.value != "1.0" {
-		t.Errorf("Invalid value. Instead of '1.0', we have '%s'.", fi.value)
-	}
-}
-func TestFormItemFloatValueToString(t *testing.T) {
-	fi := testFormItemFloat(t)
-	fi.AddCursor()
-	fi.value = "-"
-	if fi.ValueToString() != "-" {
-		t.Errorf("Invalid valuestring. instead of '-', we have '%s'.", fi.ValueToString())
-	}
-	fi.value = "-3"
-	if fi.ValueToString() != "-3" {
-		t.Errorf("Invalid valuestring. instead of '-3', we have '%s'.", fi.ValueToString())
-	}
-	fi.value = "-3."
-	if fi.ValueToString() != "-3." {
-		t.Errorf("Invalid valuestring. instead of '-3.', we have '%s'.", fi.ValueToString())
-	}
-	fi.value = "-3.3"
-	if fi.ValueToString() != "-3.3" {
-		t.Errorf("Invalid valuestring. instead of '-3.3', we have '%s'.", fi.ValueToString())
-	}
-
-}
-func TestFormItemFloatDeleteLastCharacter(t *testing.T) {
-	fi := testFormItemFloat(t)
-	fi.AddCursor()
-	fi.CharCallback('-', 0.1)
-	fi.CharCallback('3', 0.1)
-	fi.CharCallback('.', 0.1)
-	fi.CharCallback('3', 0.1)
-	if fi.ValueToString() != "-3.3" {
-		t.Errorf("Invalid valuestring. instead of '-3.3', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "NF" {
-		t.Errorf("Invalid typeState. Instead of 'NF', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-3." {
-		t.Errorf("Invalid valuestring. instead of '-3.', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "N." {
-		t.Errorf("Invalid typeState. Instead of 'N.', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-3" {
-		t.Errorf("Invalid valuestring. instead of '-3', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "NI" {
-		t.Errorf("Invalid typeState. Instead of 'NI', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-" {
-		t.Errorf("Invalid valuestring. instead of '-', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "N" {
-		t.Errorf("Invalid typeState. Instead of 'N', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "" {
-		t.Errorf("Invalid valuestring. instead of '', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('-', 0.1)
-	fi.CharCallback('0', 0.1)
-	fi.CharCallback('.', 0.1)
-	fi.CharCallback('2', 0.1)
-	if fi.ValueToString() != "-0.2" {
-		t.Errorf("Invalid valuestring. instead of '-0.2', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "NF" {
-		t.Errorf("Invalid typeState. Instead of 'NF', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-0." {
-		t.Errorf("Invalid valuestring. instead of '-0.', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "N." {
-		t.Errorf("Invalid typeState. Instead of 'N.', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-0" {
-		t.Errorf("Invalid valuestring. instead of '-0', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "N0" {
-		t.Errorf("Invalid typeState. Instead of 'N0', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-" {
-		t.Errorf("Invalid valuestring. instead of '-', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "N" {
-		t.Errorf("Invalid typeState. Instead of 'N', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "" {
-		t.Errorf("Invalid valuestring. instead of '', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('3', 0.1)
-	fi.CharCallback('.', 0.1)
-	fi.CharCallback('3', 0.1)
-	if fi.ValueToString() != "3.3" {
-		t.Errorf("Invalid valuestring. instead of '3.3', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "PF" {
-		t.Errorf("Invalid typeState. Instead of 'PF', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "3." {
-		t.Errorf("Invalid valuestring. instead of '3.', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "P." {
-		t.Errorf("Invalid typeState. Instead of 'P.', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "3" {
-		t.Errorf("Invalid valuestring. instead of '3', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "PI" {
-		t.Errorf("Invalid typeState. Instead of 'PI', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "" {
-		t.Errorf("Invalid valuestring. instead of '', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeState)
-	}
-	fi.CharCallback('0', 0.1)
-	fi.CharCallback('.', 0.1)
-	fi.CharCallback('3', 0.1)
-	if fi.ValueToString() != "0.3" {
-		t.Errorf("Invalid valuestring. instead of '0.3', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "PF" {
-		t.Errorf("Invalid typeState. Instead of 'PF', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "0." {
-		t.Errorf("Invalid valuestring. instead of '0.', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "P." {
-		t.Errorf("Invalid typeState. Instead of 'P.', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "0" {
-		t.Errorf("Invalid valuestring. instead of '0', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "P0" {
-		t.Errorf("Invalid typeState. Instead of 'P0', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "" {
-		t.Errorf("Invalid valuestring. instead of '', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeState)
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "" {
-		t.Errorf("Invalid valuestring. instead of '', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeState != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeState)
-	}
-}
-func testFormItemText(t *testing.T) *FormItemText {
-	mat := material.Chrome
-	pos := mgl32.Vec3{0, 0, 0}
-	fi := NewFormItemText(DefaultMaxWidth, ITEM_WIDTH_HALF, DefaultFormItemLabel, DefaultFormItemDescription, mat, pos, wrapperMock)
-
-	if fi.label != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.label)
-	}
-	if fi.description != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.description)
-	}
-	return fi
-}
-func TestNewFormItemText(t *testing.T) {
-	_ = testFormItemText(t)
-}
-func TestFormItemTextGetLabel(t *testing.T) {
-	fi := testFormItemText(t)
-	if fi.GetLabel() != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.GetLabel())
-	}
-}
-func TestFormItemTextGetDescription(t *testing.T) {
-	fi := testFormItemText(t)
-	if fi.GetDescription() != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.GetDescription())
-	}
-}
-func TestFormItemTextGetValue(t *testing.T) {
-	fi := testFormItemText(t)
-	val := "test value"
-	fi.value = val
-	if fi.GetValue() != val {
-		t.Errorf("Invalid form item value. Instead of '%s', it is '%s'.", val, fi.GetValue())
-	}
-}
-func TestFormItemTextGetSurface(t *testing.T) {
-	fi := testFormItemText(t)
-	if fi.GetSurface() != fi.meshes[0] {
-		t.Error("Invalid surface mesh")
-	}
-}
-func TestFormItemTextGetTarget(t *testing.T) {
-	fi := testFormItemText(t)
-	if fi.GetTarget() != fi.meshes[1] {
-		t.Error("Invalid target mesh")
-	}
-}
-func TestFormItemTextAddCursor(t *testing.T) {
-	fi := testFormItemText(t)
-	if len(fi.meshes) != 2 {
-		t.Error("Invalid number of target mesh")
-	}
-	fi.AddCursor()
-	if len(fi.meshes) != 3 {
-		t.Error("Invalid number of target mesh")
-	}
-}
-func TestFormItemTextDeleteCursor(t *testing.T) {
-	fi := testFormItemText(t)
-	fi.AddCursor()
-	if len(fi.meshes) != 3 {
-		t.Error("Invalid number of target mesh")
-	}
-	fi.DeleteCursor()
-	if len(fi.meshes) != 2 {
-		t.Error("Invalid number of target mesh")
-	}
-}
-func TestFormItemTextCharCallback(t *testing.T) {
-	fi := testFormItemText(t)
-	fi.AddCursor()
-	fi.CharCallback('b', 0.1)
-	if fi.value != "b" {
-		t.Errorf("Invalid value. Instead of 'b', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('a', 0.1)
-	if fi.value != "ba" {
-		t.Errorf("Invalid value. Instead of 'ba', we have '%s'.", fi.value)
-	}
-	fi.CharCallback(' ', 0.1)
-	if fi.value != "ba " {
-		t.Errorf("Invalid value. Instead of 'ba ', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('b', 0.1)
-	if fi.value != "ba b" {
-		t.Errorf("Invalid value. Instead of 'ba b', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('b', 0.1)
-	if fi.value != "ba bb" {
-		t.Errorf("Invalid value. Instead of 'ba bb', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('b', 0.1)
-	if fi.value != "ba bbb" {
-		t.Errorf("Invalid value. Instead of 'ba bbb', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('b', 0.1)
-	if fi.value != "ba bbbb" {
-		t.Errorf("Invalid value. Instead of 'ba bbbb', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('b', 0.1)
-	if fi.value != "ba bbbbb" {
-		t.Errorf("Invalid value. Instead of 'ba bbbbb', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('b', 0.1)
-	if fi.value != "ba bbbbbb" {
-		t.Errorf("Invalid value. Instead of 'ba bbbbbb', we have '%s'.", fi.value)
-	}
-	valueBase := "ba bbbbbb"
-	for i := 0; i < 25; i++ {
-		valueBase = valueBase + "b"
-		fi.CharCallback('b', 0.1)
-		if fi.value != valueBase {
-			t.Errorf("Invalid value (%d). Instead of '%s', we have '%s'.", i, valueBase, fi.value)
-		}
-	}
-	for i := 0; i < 10; i++ {
-		fi.CharCallback('b', 0.1)
-		if fi.value != valueBase {
-			t.Errorf("Invalid value (%d). Instead of '%s', we have '%s'.", i, valueBase, fi.value)
-		}
-	}
-	fi = testFormItemText(t)
-	fi.AddCursor()
-	validator := func(s string) bool {
-		return s != "not"
-	}
-	fi.SetValidator(validator)
-	fi.CharCallback('n', 0.1)
-	if fi.value != "n" {
-		t.Errorf("Invalid value. Instead of 'n', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('o', 0.1)
-	if fi.value != "no" {
-		t.Errorf("Invalid value. Instead of 'no', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('t', 0.1)
-	if fi.value != "no" {
-		t.Errorf("Invalid value. Instead of 'no', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('r', 0.1)
-	if fi.value != "nor" {
-		t.Errorf("Invalid value. Instead of 'nor', we have '%s'.", fi.value)
-	}
-}
-func TestFormItemTextValueToString(t *testing.T) {
-	fi := testFormItemText(t)
-	val := fi.ValueToString()
-	if val != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", val)
-	}
-	fi.value = "test value"
-	val = fi.ValueToString()
-	if val != "test value" {
-		t.Errorf("Invalid value. Instead of 'test value', we have '%s'.", val)
-	}
-}
-func TestFormItemTextDeleteLastCharacter(t *testing.T) {
-	fi := testFormItemText(t)
-	fi.value = "12345"
-	fi.charOffsets = []float32{0.1, 0.1, 0.1, 0.1, 0.1}
-	fi.DeleteLastCharacter()
-	if fi.value != "1234" {
-		t.Errorf("Invalid value. Instead of '1234', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "123" {
-		t.Errorf("Invalid value. Instead of '123', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "12" {
-		t.Errorf("Invalid value. Instead of '12', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-}
-func testFormItemInt64(t *testing.T) *FormItemInt64 {
-	mat := material.Chrome
-	pos := mgl32.Vec3{0, 0, 0}
-	fi := NewFormItemInt64(DefaultMaxWidth, ITEM_WIDTH_HALF, DefaultFormItemLabel, DefaultFormItemDescription, mat, pos, wrapperMock)
-
-	if fi.label != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.label)
-	}
-	if fi.description != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.description)
-	}
-	return fi
-}
-func TestNewFormItemInt64(t *testing.T) {
-	_ = testFormItemInt64(t)
-}
-func TestFormItemInt64GetLabel(t *testing.T) {
-	fi := testFormItemInt64(t)
-	if fi.GetLabel() != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.GetLabel())
-	}
-}
-func TestFormItemInt64GetDescription(t *testing.T) {
-	fi := testFormItemInt64(t)
-	if fi.GetDescription() != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.GetDescription())
-	}
-}
-func TestFormItemInt64GetValue(t *testing.T) {
-	fi := testFormItemInt64(t)
-	val := int64(3)
-	fi.value = "3"
-	if fi.GetValue() != val {
-		t.Errorf("Invalid form item value. Instead of '%d', it is '%d'.", val, fi.GetValue())
-	}
-}
-func TestFormItemInt64GetSurface(t *testing.T) {
-	fi := testFormItemInt64(t)
-	if fi.GetSurface() != fi.meshes[0] {
-		t.Error("Invalid surface mesh")
-	}
-}
-func TestFormItemInt64GetTarget(t *testing.T) {
-	fi := testFormItemInt64(t)
-	if fi.GetTarget() != fi.meshes[1] {
-		t.Error("Invalid target mesh")
-	}
-}
-func TestFormItemInt64AddCursor(t *testing.T) {
-	fi := testFormItemInt64(t)
-	if len(fi.meshes) != 2 {
-		t.Error("Invalid number of target mesh")
-	}
-	fi.AddCursor()
-	if len(fi.meshes) != 3 {
-		t.Error("Invalid number of target mesh")
-	}
-}
-func TestFormItemInt64DeleteCursor(t *testing.T) {
-	fi := testFormItemInt64(t)
-	fi.AddCursor()
-	if len(fi.meshes) != 3 {
-		t.Error("Invalid number of target mesh")
-	}
-	fi.DeleteCursor()
-	if len(fi.meshes) != 2 {
-		t.Error("Invalid number of target mesh")
-	}
-}
-func TestFormItemInt64CharCallback(t *testing.T) {
-	fi := testFormItemInt64(t)
-	fi.AddCursor()
-	fi.DeleteLastCharacter()
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('b', 0.1)
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.value != "0" {
-		t.Errorf("Invalid value. Instead of '0', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	fi.CharCallback('-', 0.1)
-	if fi.value != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "-1" {
-		t.Errorf("Invalid value. Instead of '-1', we have '%s'.", fi.value)
-	}
-	fi = testFormItemInt64(t)
-	fi.AddCursor()
-	fi.CharCallback('1', 0.1)
-	if fi.value != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	fi.value = strconv.Itoa(math.MaxInt64 - 10)
-	maxVal := strconv.Itoa(math.MaxInt64-10) + "1"
-	fi.CharCallback('1', 0.1)
-	if fi.value != maxVal {
-		t.Errorf("Invalid value. Instead of '%s', we have '%s'.", maxVal, fi.value)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != maxVal {
-		t.Errorf("Invalid value. Instead of '%s', we have '%s'.", maxVal, fi.value)
-	}
-	fi = testFormItemInt64(t)
-	fi.AddCursor()
-	validator := func(i int64) bool {
-		return i >= -300 && i <= 200
-	}
-	fi.SetValidator(validator)
-	fi.CharCallback('1', 0.1)
-	if fi.value != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "11" {
-		t.Errorf("Invalid value. Instead of '11', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "111" {
-		t.Errorf("Invalid value. Instead of '111', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.value != "111" {
-		t.Errorf("Invalid value. Instead of '111', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	fi.DeleteLastCharacter()
-	fi.DeleteLastCharacter()
-	fi.CharCallback('-', 0.1)
-	fi.CharCallback('3', 0.1)
-	if fi.value != "-3" {
-		t.Errorf("Invalid value. Instead of '-3', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('3', 0.1)
-	if fi.value != "-33" {
-		t.Errorf("Invalid value. Instead of '-33', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('3', 0.1)
-	if fi.value != "-33" {
-		t.Errorf("Invalid value. Instead of '-33', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	fi.CharCallback('0', 0.1)
-	if fi.value != "-30" {
-		t.Errorf("Invalid value. Instead of '-30', we have '%s'.", fi.value)
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.value != "-300" {
-		t.Errorf("Invalid value. Instead of '-300', we have '%s'.", fi.value)
-	}
-}
-func TestFormItemInt64ValueToString(t *testing.T) {
-	fi := testFormItemInt64(t)
-	val := fi.ValueToString()
-	if val != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", val)
-	}
-	fi.value = "3"
-	val = fi.ValueToString()
-	if val != "3" {
-		t.Errorf("Invalid value. Instead of '3', we have '%s'.", val)
-	}
-}
-func TestFormItemInt64DeleteLastCharacter(t *testing.T) {
-	fi := testFormItemInt64(t)
-	fi.value = "12345"
-	fi.typeState = "PI"
-	fi.charOffsets = []float32{0.1, 0.1, 0.1, 0.1, 0.1}
-	fi.DeleteLastCharacter()
-	if fi.value != "1234" {
-		t.Errorf("Invalid value. Instead of '1234', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "123" {
-		t.Errorf("Invalid value. Instead of '123', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "12" {
-		t.Errorf("Invalid value. Instead of '12', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-	fi.value = "-12"
-	fi.typeState = "NI"
-	fi.charOffsets = []float32{0.1, 0.1, 0.1}
-	fi.DeleteLastCharacter()
-	if fi.value != "-1" {
-		t.Errorf("Invalid value. Instead of '-1', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-	fi.DeleteLastCharacter()
-	if fi.value != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.value)
-	}
-}
-
-func testFormItemVector(t *testing.T) *FormItemVector {
-	mat := material.Chrome
-	pos := mgl32.Vec3{0, 0, 0}
-	fi := NewFormItemVector(DefaultMaxWidth, ITEM_WIDTH_FULL, DefaultFormItemLabel, DefaultFormItemDescription, 10, mat, pos, wrapperMock)
-
-	if fi.label != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.label)
-	}
-	if fi.description != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.description)
-	}
-	return fi
-}
-func TestNewFormItemVector(t *testing.T) {
-	_ = testFormItemVector(t)
-}
-func TestFormItemVectorGetLabel(t *testing.T) {
-	fi := testFormItemVector(t)
-	if fi.GetLabel() != DefaultFormItemLabel {
-		t.Errorf("Invalid form item label. Instead of '%s', we have '%s'.", DefaultFormItemLabel, fi.GetLabel())
-	}
-}
-func TestFormItemVectorGetDescription(t *testing.T) {
-	fi := testFormItemVector(t)
-	if fi.GetDescription() != DefaultFormItemDescription {
-		t.Errorf("Invalid form item description. Instead of '%s', we have '%s'.", DefaultFormItemDescription, fi.GetDescription())
-	}
-}
-func TestFormItemVectorGetValue(t *testing.T) {
-	fi := testFormItemVector(t)
-	nullVec := mgl32.Vec3{0, 0, 0}
-	if fi.GetValue() != nullVec {
-		t.Errorf("Invalid form item value. Instead of '%v', it is '%v'.", nullVec, fi.GetValue())
-	}
-	fi.values[0] = "3.2"
-	fi.values[1] = "4.1"
-	fi.values[2] = "5.0"
-	vec := mgl32.Vec3{3.2, 4.1, 5.0}
-	if fi.GetValue() != vec {
-		t.Errorf("Invalid form item value. Instead of '%v', it is '%v'.", vec, fi.GetValue())
-	}
-}
-func TestFormItemVectorGetSurface(t *testing.T) {
-	fi := testFormItemVector(t)
-	if fi.GetSurface() != fi.meshes[0] {
-		t.Error("Invalid surface mesh")
-	}
-}
-func TestFormItemVectorGetTarget(t *testing.T) {
-	fi := testFormItemVector(t)
-	if fi.GetTarget() != fi.meshes[1] {
-		t.Error("Invalid target mesh")
-	}
-}
-func TestFormItemVectorAddCursor(t *testing.T) {
-	fi := testFormItemVector(t)
-	if len(fi.meshes) != 4 {
-		t.Error("Invalid number of target mesh")
-	}
-	fi.AddCursor()
-	if len(fi.meshes) != 5 {
-		t.Error("Invalid number of target mesh")
-	}
-}
-func TestFormItemVectorDeleteCursor(t *testing.T) {
-	fi := testFormItemVector(t)
-	fi.AddCursor()
-	if len(fi.meshes) != 5 {
-		t.Error("Invalid number of target mesh")
-	}
-	fi.DeleteCursor()
-	if len(fi.meshes) != 4 {
-		t.Error("Invalid number of target mesh")
-	}
-}
-func TestFormItemVectorCharCallback(t *testing.T) {
-	fi := testFormItemVector(t)
-	fi.AddCursor()
-	// start with 0
-	fi.CharCallback('0', 0.1)
-	if fi.values[0] != "0" {
-		t.Errorf("Invalid value. Instead of '0', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "P0" {
-		t.Errorf("Invalid typeState. Instead of 'P0', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.values[0] != "0" {
-		t.Errorf("Invalid value. Instead of '0', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "P0" {
-		t.Errorf("Invalid typeState. Instead of 'P0', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.values[0] != "0." {
-		t.Errorf("Invalid value. Instead of '0.', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "P." {
-		t.Errorf("Invalid typeState. Instead of 'P.', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('2', 0.1)
-	if fi.values[0] != "0.2" {
-		t.Errorf("Invalid value. Instead of '0.2', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "PF" {
-		t.Errorf("Invalid typeState. Instead of 'PF', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('8', 0.1)
-	if fi.values[0] != "0.28" {
-		t.Errorf("Invalid value. Instead of '0.28', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "PF" {
-		t.Errorf("Invalid typeState. Instead of 'PF', we have '%s'.", fi.typeStates[0])
-	}
-	// start with .
-	fi = testFormItemVector(t)
-	fi.AddCursor()
-	fi.CharCallback('.', 0.1)
-	if fi.values[0] != "" {
-		t.Errorf("Invalid value. Instead of '', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.values[0] != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "PI" {
-		t.Errorf("Invalid typeState. Instead of 'PI', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('2', 0.1)
-	if fi.values[0] != "12" {
-		t.Errorf("Invalid value. Instead of '12', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "PI" {
-		t.Errorf("Invalid typeState. Instead of 'PI', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.values[0] != "12." {
-		t.Errorf("Invalid value. Instead of '12.', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "P." {
-		t.Errorf("Invalid typeState. Instead of 'P.', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('8', 0.1)
-	if fi.values[0] != "12.8" {
-		t.Errorf("Invalid value. Instead of '12.8', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "PF" {
-		t.Errorf("Invalid typeState. Instead of 'PF', we have '%s'.", fi.typeStates[0])
-	}
-	// start with -
-	fi = testFormItemVector(t)
-	fi.AddCursor()
-	fi.CharCallback('-', 0.1)
-	if fi.values[0] != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "N" {
-		t.Errorf("Invalid typeState. Instead of 'N', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.values[0] != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "N" {
-		t.Errorf("Invalid typeState. Instead of 'N', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('1', 0.1)
-	if fi.values[0] != "-1" {
-		t.Errorf("Invalid value. Instead of '-1', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "NI" {
-		t.Errorf("Invalid typeState. Instead of 'NI', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('2', 0.1)
-	if fi.values[0] != "-12" {
-		t.Errorf("Invalid value. Instead of '-12', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "NI" {
-		t.Errorf("Invalid typeState. Instead of 'NI', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.values[0] != "-12." {
-		t.Errorf("Invalid value. Instead of '-12.', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "N." {
-		t.Errorf("Invalid typeState. Instead of 'N.', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('8', 0.1)
-	if fi.values[0] != "-12.8" {
-		t.Errorf("Invalid value. Instead of '-12.8', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "NF" {
-		t.Errorf("Invalid typeState. Instead of 'NF', we have '%s'.", fi.typeStates[0])
-	}
-	// start with -0.
-	fi = testFormItemVector(t)
-	fi.AddCursor()
-	fi.CharCallback('-', 0.1)
-	if fi.values[0] != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "N" {
-		t.Errorf("Invalid typeState. Instead of 'N', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.values[0] != "-0" {
-		t.Errorf("Invalid value. Instead of '-0', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "N0" {
-		t.Errorf("Invalid typeState. Instead of 'N0', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.values[0] != "-0." {
-		t.Errorf("Invalid value. Instead of '-0.', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "N." {
-		t.Errorf("Invalid typeState. Instead of 'N.', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('8', 0.1)
-	if fi.values[0] != "-0.8" {
-		t.Errorf("Invalid value. Instead of '-0.8', we have '%s'.", fi.values[0])
-	}
-	if fi.typeStates[0] != "NF" {
-		t.Errorf("Invalid typeState. Instead of 'NF', we have '%s'.", fi.typeStates[0])
-	}
-	// validator
-	fi = testFormItemVector(t)
-	fi.AddCursor()
-	validator := func(i float32) bool {
-		return i <= 1.0 && i >= 0.0
-	}
-	fi.SetValidator(validator)
-	fi.CharCallback('-', 0.1)
-	if fi.values[0] != "-" {
-		t.Errorf("Invalid value. Instead of '-', we have '%s'.", fi.values[0])
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.values[0] != "-0" {
-		t.Errorf("Invalid value. Instead of '-0', we have '%s'.", fi.values[0])
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.values[0] != "-0." {
-		t.Errorf("Invalid value. Instead of '-0.', we have '%s'.", fi.values[0])
-	}
-	fi.CharCallback('3', 0.1)
-	if fi.values[0] != "-0." {
-		t.Errorf("Invalid value. Instead of '-0.', we have '%s'.", fi.values[0])
-	}
-	fi.SetTarget(1)
-	fi.CharCallback('1', 0.1)
-	if fi.values[1] != "1" {
-		t.Errorf("Invalid value. Instead of '1', we have '%s'.", fi.values[0])
-	}
-	fi.CharCallback('.', 0.1)
-	if fi.values[1] != "1." {
-		t.Errorf("Invalid value. Instead of '1.', we have '%s'.", fi.values[0])
-	}
-	fi.CharCallback('3', 0.1)
-	if fi.values[1] != "1." {
-		t.Errorf("Invalid value. Instead of '1.', we have '%s'.", fi.values[1])
-	}
-	fi.CharCallback('0', 0.1)
-	if fi.values[1] != "1.0" {
-		t.Errorf("Invalid value. Instead of '1.0', we have '%s'.", fi.values[1])
-	}
-}
-func TestFormItemVectorValueToString(t *testing.T) {
-	fi := testFormItemVector(t)
-	fi.AddCursor()
-	fi.values[0] = "-"
-	if fi.ValueToString() != "-" {
-		t.Errorf("Invalid valuestring. instead of '-', we have '%s'.", fi.ValueToString())
-	}
-	fi.values[0] = "-3"
-	if fi.ValueToString() != "-3" {
-		t.Errorf("Invalid valuestring. instead of '-3', we have '%s'.", fi.ValueToString())
-	}
-	fi.values[0] = "-3."
-	if fi.ValueToString() != "-3." {
-		t.Errorf("Invalid valuestring. instead of '-3.', we have '%s'.", fi.ValueToString())
-	}
-	fi.values[0] = "-3.3"
-	if fi.ValueToString() != "-3.3" {
-		t.Errorf("Invalid valuestring. instead of '-3.3', we have '%s'.", fi.ValueToString())
-	}
-}
-func TestFormItemVectorDeleteLastCharacter(t *testing.T) {
-	fi := testFormItemVector(t)
-	fi.AddCursor()
-	fi.CharCallback('-', 0.1)
-	fi.CharCallback('3', 0.1)
-	fi.CharCallback('.', 0.1)
-	fi.CharCallback('3', 0.1)
-	if fi.ValueToString() != "-3.3" {
-		t.Errorf("Invalid valuestring. instead of '-3.3', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "NF" {
-		t.Errorf("Invalid typeState. Instead of 'NF', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-3." {
-		t.Errorf("Invalid valuestring. instead of '-3.', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "N." {
-		t.Errorf("Invalid typeState. Instead of 'N.', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-3" {
-		t.Errorf("Invalid valuestring. instead of '-3', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "NI" {
-		t.Errorf("Invalid typeState. Instead of 'NI', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-" {
-		t.Errorf("Invalid valuestring. instead of '-', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "N" {
-		t.Errorf("Invalid typeState. Instead of 'N', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "" {
-		t.Errorf("Invalid valuestring. instead of '', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('-', 0.1)
-	fi.CharCallback('0', 0.1)
-	fi.CharCallback('.', 0.1)
-	fi.CharCallback('2', 0.1)
-	if fi.ValueToString() != "-0.2" {
-		t.Errorf("Invalid valuestring. instead of '-0.2', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "NF" {
-		t.Errorf("Invalid typeState. Instead of 'NF', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-0." {
-		t.Errorf("Invalid valuestring. instead of '-0.', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "N." {
-		t.Errorf("Invalid typeState. Instead of 'N.', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-0" {
-		t.Errorf("Invalid valuestring. instead of '-0', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "N0" {
-		t.Errorf("Invalid typeState. Instead of 'N0', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "-" {
-		t.Errorf("Invalid valuestring. instead of '-', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "N" {
-		t.Errorf("Invalid typeState. Instead of 'N', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "" {
-		t.Errorf("Invalid valuestring. instead of '', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('3', 0.1)
-	fi.CharCallback('.', 0.1)
-	fi.CharCallback('3', 0.1)
-	if fi.ValueToString() != "3.3" {
-		t.Errorf("Invalid valuestring. instead of '3.3', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "PF" {
-		t.Errorf("Invalid typeState. Instead of 'PF', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "3." {
-		t.Errorf("Invalid valuestring. instead of '3.', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "P." {
-		t.Errorf("Invalid typeState. Instead of 'P.', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "3" {
-		t.Errorf("Invalid valuestring. instead of '3', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "PI" {
-		t.Errorf("Invalid typeState. Instead of 'PI', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "" {
-		t.Errorf("Invalid valuestring. instead of '', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeStates[0])
-	}
-	fi.CharCallback('0', 0.1)
-	fi.CharCallback('.', 0.1)
-	fi.CharCallback('3', 0.1)
-	if fi.ValueToString() != "0.3" {
-		t.Errorf("Invalid valuestring. instead of '0.3', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "PF" {
-		t.Errorf("Invalid typeState. Instead of 'PF', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "0." {
-		t.Errorf("Invalid valuestring. instead of '0.', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "P." {
-		t.Errorf("Invalid typeState. Instead of 'P.', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "0" {
-		t.Errorf("Invalid valuestring. instead of '0', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "P0" {
-		t.Errorf("Invalid typeState. Instead of 'P0', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "" {
-		t.Errorf("Invalid valuestring. instead of '', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeStates[0])
-	}
-	fi.DeleteLastCharacter()
-	if fi.ValueToString() != "" {
-		t.Errorf("Invalid valuestring. instead of '', we have '%s'.", fi.ValueToString())
-	}
-	if fi.typeStates[0] != "P" {
-		t.Errorf("Invalid typeState. Instead of 'P', we have '%s'.", fi.typeStates[0])
 	}
 }
